requirements.txt
# RunPod WhisperX 半自动种子识别说话人日志系统
# 统一依赖文件 - 稳定版本组合，避免版本冲突

# ============================================================================
# PyTorch生态系统 (固定兼容版本 - 重要!)
# ============================================================================
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0

# ============================================================================
# 核心数据科学库 (更新以兼容WhisperX 3.3.4)
# ============================================================================
numpy>=2.0.2
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0

# ============================================================================
# 音频处理
# ============================================================================
librosa>=0.10.0,<1.0.0
soundfile>=0.12.0,<1.0.0

# ============================================================================
# 自然语言处理和机器学习
# ============================================================================
transformers>=4.35.0,<5.0.0
accelerate>=0.20.0,<1.0.0

# ============================================================================
# 说话人识别 (固定版本确保稳定性)
# ============================================================================
pyannote.audio==3.1.1

# ============================================================================
# WhisperX (半自动种子识别系统核心) - 固定版本避免依赖冲突
# ============================================================================
whisperx==3.3.4

# ============================================================================
# 工具库
# ============================================================================
tqdm>=4.65.0
omegaconf>=2.3.0

# ============================================================================
# 安装说明
# ============================================================================
# 1. 安装依赖: pip install -r requirements.txt
# 2. 设置Token: export HF_TOKEN='your_huggingface_token'
# 3. 运行脚本: python process_audio_large.py
#
# 注意: 如果遇到版本冲突，运行 python setup_environment.py 自动修复
